XRDDEV-1715 Move /user api and model to the common module
