+++ src/proxy-ui-api/frontend/src/store/modules/client.ts
@@ -26,7 +26,11 @@
 import axios from 'axios';
 import { ActionTree, GetterTree, Module, MutationTree } from 'vuex';
 import { RootState } from '../types';
-import { CertificateDetails, Client, TokenCertificate } from '@/openapi-types/ss-types';
+import {
+  CertificateDetails,
+  Client,
+  TokenCertificate,
+} from '@/openapi-types/ss-types';
 import { encodePathParameter } from '@/util/api';
 
 export interface ClientState {
