+++ src/proxy-ui-api/frontend/src/views/Clients/Services/Services.vue
@@ -211,7 +211,11 @@ import DisableServiceDescDialog from './DisableServiceDescDialog.vue';
 import WarningDialog from '@/components/service/WarningDialog.vue';
 import ServiceIcon from '@/components/ui/ServiceIcon.vue';
 
-import { Service, ServiceDescription, ServiceType } from '@/openapi-types/ss-types';
+import {
+  Service,
+  ServiceDescription,
+  ServiceType,
+} from '@/openapi-types/ss-types';
 import { ServiceTypeEnum } from '@/domain';
 import { Prop } from 'vue/types/options';
 import { sortServiceDescriptionServices } from '@/util/sorting';
