+++ src/proxy-ui-api/src/main/resources/openapi-definition.yaml
@@ -2893,32 +2893,6 @@ paths:
           description: request specified an invalid format
         '500':
           description: internal server error
-  /user:
-    get:
-      tags:
-        - user
-      summary: get user data for the logged user
-      operationId: getUser
-      description: <h3>Administrator gets user data from backend.</h3>
-      responses:
-        '200':
-          description: user details
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/User'
-        '400':
-          description: request was invalid
-        '401':
-          description: authentication credentials are missing
-        '403':
-          description: request has been refused
-        '404':
-          description: resource requested does not exists
-        '406':
-          description: request specified an invalid format
-        '500':
-          description: internal server error
   /security-servers:
     get:
       tags:
@@ -6829,35 +6803,6 @@ components:
       enum:
         - SOFTWARE
         - HARDWARE
-    User:
-      type: object
-      description: x-road user
-      required:
-        - username
-        - roles
-        - permissions
-      properties:
-        username:
-          type: string
-          format: text
-          description: user username
-          example: Guest
-          minLength: 1
-          maxLength: 255
-        roles:
-          type: array
-          description: user roles
-          uniqueItems: true
-          items:
-            type: string
-            format: text
-        permissions:
-          type: array
-          description: user permissions
-          uniqueItems: true
-          items:
-            type: string
-            format: text
     VersionInfo:
       type: object
       description: version information
